% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\docType{data}
\name{demographic}
\alias{demographic}
\title{DIVINE's demographic database}
\format{
A data frame with 5813 rows and 6 columns
\describe{
\item{record_id:}{Identifier of each record. This information does not match the real data.}
\item{covid_wave:}{COVID-19 wave}
\item{center:}{A factor with levels \verb{Hospital A}, \verb{Hospital B}, \verb{Hospital C}, \verb{Hospital D}, and             \verb{Hospital E}. Center of admission}
\item{sex:}{A factor with levels \code{Male} and \code{Female}. Sex at birth}
\item{age:}{Age at hospital admission}
\item{race:}{A factor with levels \code{Asian}, \code{Caucasian}, \code{Latina}, \code{Black}, and \code{Others}. Race}
}
}
\usage{
data(demographic)
}
\description{
Demographic data of patients included in the DIVINE cohort. Data was collected at hospital admission.
}
\references{
Pallarès, N., Tebé, C., Abelenda-Alonso, G., Rombauts, A., Oriol, I., Simonetti, A. F., Rodríguez-Molinero, A., Izquierdo, E., Díaz-Brito, V., Molist, G., Gómez Melis, G., Carratalà, J., Videla, S., & MetroSud and Divine study groups (2023). Characteristics and Outcomes by Ceiling of Care of Subjects Hospitalized with COVID-19 During Four Waves of the Pandemic in a Metropolitan Area: A Multicenter Cohort Study. Infectious diseases and therapy, 12(1), 273–289. https://doi.org/10.1007/s40121-022-00705-w
}
\keyword{datasets}
