% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\docType{data}
\name{treatment}
\alias{treatment}
\title{DIVINE's database on treatments previous to hospital admission}
\format{
A data frame with 5813 rows and 11 columns
\describe{
\item{record_id:}{Identifier of each record. This information does not match the real data.}
\item{covid_wave:}{COVID-19 wave}
\item{center:}{A factor with levels \verb{Hospital A}, \verb{Hospital B}, \verb{Hospital C}, \verb{Hospital D}, and             \verb{Hospital E}. Center of admission}
\item{statins_pre:}{Has the patient }
\item{cortis_pre:}{Previous treatment with corticosteroids}
\item{acei_pre:}{Previous treatment with angiotensin-converting enzyme (ACE) inhibitors}
\item{ara2_pre:}{Previous treatment with angiotensin II receptor antagonists (ARA-II)}
\item{cortis_systemic_pre:}{Routine treatment with systemic corticosteroids}
\item{cortis_inhaled_pre:}{Routine treatment with inhaled corticosteroids}
\item{anticoagulants_pre:}{Previous treatment with anticoagulants}
\item{immunosuppre_pre:}{Usual treatment with immunosuppressants}
}
}
\usage{
data(treatment)
}
\description{
Information on previous treatments for patients included in the DIVINE cohort. Data was collected at hospital admission.
}
\references{
Pallarès, N., Tebé, C., Abelenda-Alonso, G., Rombauts, A., Oriol, I., Simonetti, A. F., Rodríguez-Molinero, A., Izquierdo, E., Díaz-Brito, V., Molist, G., Gómez Melis, G., Carratalà, J., Videla, S., & MetroSud and Divine study groups (2023). Characteristics and Outcomes by Ceiling of Care of Subjects Hospitalized with COVID-19 During Four Waves of the Pandemic in a Metropolitan Area: A Multicenter Cohort Study. Infectious diseases and therapy, 12(1), 273–289. https://doi.org/10.1007/s40121-022-00705-w
}
\keyword{datasets}
