% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidities.R
\docType{data}
\name{comorbidities}
\alias{comorbidities}
\title{DIVINE's demographic database}
\format{
A data frame with 5813 rows and 43 columns
\describe{
\item{record_id:}{Identifier of each record. This information does not match the real data.}
\item{covid_wave:}{COVID-19 wave}
\item{center:}{A factor with levels \verb{Hospital A}, \verb{Hospital B}, \verb{Hospital C}, \verb{Hospital D}, and             \verb{Hospital E}. Center of admission}
\item{smoker:}{A factor with levels \code{Ex-smoker}, \code{No} and \code{Yes}. Smoking status}
\item{alcohol:}{Consumption of alcohol}
\item{residence_center:}{Is the patient currently living in a long-term facility?}
\item{sociofunctional:}{XXX NECESSARI COMENTAR-LA. Sociofunctional status}
\item{chronicity_degree:}{A factor with levels \code{No}, \code{MACA} and \code{PCC}. Is the patient a chronic             complex patient (PCC) or a patient with advanced chronic disease (MACA)?}
\item{barthel_score:}{Punctuation in the Barthel scale used to measure performance in activities #'          of daily living}
\item{weight:}{Weight (kg)}
\item{height:}{Height (cm)}
\item{body mass index:}{Body mass index computed as weight (kg)/height(m)^2}
\item{obesity:}{Obesity}
\item{dm:}{Diabetes mellitus Type 2}
\item{type_dm:}{A factor with levels \verb{With target organ involvement} and \verb{Without complications}.           For patients with diabetes mellitus type 2, type of disease}
\item{chronic_lung_disease:}{Chronic lung disease (including COPD, asthma and obstructive sleep #'          apnea, among others)}
\item{chronic_kidney_disease:}{Severe chronic kidney disease}
\item{renal_therapy:}{Is the patient currently receiving renal replacement therapy?}
\item{heart_disease:}{Heart failure}
\item{coronary_disease:}{Coronary heart disease}
\item{myocardial_infarction:}{Has the patient ever had a heart attack?}
\item{hematologic_neo:}{Haematological neoplasia}
\item{hematologic_neo_type:}{A factor with levels \code{Leukemia}, \code{Lymphoma} and \code{Myeloma}. For                 patients with Haematological neoplasia, type of disease.}
\item{acute_leukemia:}{Acute leukemia}
\item{malignant_hemopathy:}{MMalignant hemopathy}
\item{non_metastatic_neo:}{Non-Metastatic Neoplasia}
\item{metastatic_neo:}{Metastatic Neoplasia}
\item{stroke_tia:}{Has the patient ever had a stroke or a transient ischemic attack?}
\item{peripheral_vasculopathy:}{Peripheral artery disease}
\item{dementia:}{Dementia}
\item{mild_liver_disease:}{Mild liver disease}
\item{severe_liver_disease:}{Severe liver disease}
\item{connective_tissue_disease:}{Connective tissue disease}
\item{peptic_ulcer:}{Peptic ulcer}
\item{hemiplegia:}{Hemiplegia}
\item{hiv:}{human immunodeficiency virus}
\item{charlson_index:}{Value of the Charlson Comorbidity Index. This index predicts the ten-year         mortality for a patient given the information of their comorbid conditions}
\item{hypertension:}{Hypertension}
\item{dyslipidemia:}{Dyslipidemia }
\item{depression:}{Depression}
\item{mild_kidney_disease:}{Mild kidney disease}
\item{ceiling:}{A factor with levels \verb{Oxygen mask} (non-rebreather oxygen mask), \verb{HFNC or NIMV}         (high-flow nasal cannula and non-invasive mechanical ventilation) and \verb{IMV and ICU admission}         (invasive mechanical ventilation and acces to intensive care unit). Therapeutic ceiling of           care assigned to the patient}
\item{ceiling_dico:}{Dichotomization of variable ceiling in two levels \code{No} (\verb{IMV and ICU               admission}) and \code{Yes} (\verb{Oxygen mask} and \verb{HFNC or NIMV})}
}
}
\usage{
data(demographic)
}
\description{
Information about comorbidities of patients included in the DIVINE cohort. Data was collected at hospital admission.
#'
}
\references{
Pallarès, N., Tebé, C., Abelenda-Alonso, G., Rombauts, A., Oriol, I., Simonetti, A. F., Rodríguez-Molinero, A., Izquierdo, E., Díaz-Brito, V., Molist, G., Gómez Melis, G., Carratalà, J., Videla, S., & MetroSud and Divine study groups (2023). Characteristics and Outcomes by Ceiling of Care of Subjects Hospitalized with COVID-19 During Four Waves of the Pandemic in a Metropolitan Area: A Multicenter Cohort Study. Infectious diseases and therapy, 12(1), 273–289. https://doi.org/10.1007/s40121-022-00705-w
}
\keyword{datasets}
